requirements.txt
streamlit==1.28.1
qdrant-client==1.16.2
pillow==10.1.0
open-clip-torch==2.20.0
langchain==0.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
tqdm
